# api/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rosslib API
  description: |
    Rosslib is a Goodreads alternative. This API powers the webapp and is available
    for building CLIs, integrations, and third-party tools.

    ## Authentication
    Most read endpoints are public. Write endpoints require a JWT token obtained
    via `/auth/register` or `/auth/login`. Pass the token as:
    ```
    Authorization: Bearer <token>
    ```
    Tokens expire after 30 days.

    ## Conventions
    - Book IDs use bare Open Library work IDs (e.g. `OL82592W`, not `/works/OL82592W`)
    - Dates are RFC 3339 format (e.g. `2025-06-15T00:00:00Z`)
    - Nullable fields return `null` when unset
    - Success responses for mutations return `{"ok": true}`
    - Error responses return `{"error": "description"}`
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    ShelfDetailBook:
      type: object
      properties:
        book_id:
          type: string
          format: uuid
        open_library_id:
          type: string
        title:
          type: string
        cover_url:
          type: string
          nullable: true
        added_at:
          type: string
          format: date-time
        rating:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
...
    BookTagResponse:
      type: object
      properties:
        key_id:
          type: string
          format: uuid
        key_name:
          type: string
        key_slug:
          type: string
        value_id:
          type: string
          format: uuid
        value_name:
          type: string
        value_slug:
          type: string
...
paths:
...
  /users/{username}/tags/{path}:
    get:
      summary: Get books tagged with a path
      description: "Returns books matching the tag path or any sub-tag. E.g. /users/alice/tags/sci-fi includes books tagged sci-fi/moon."
      tags: [Collections]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: Tag path (e.g. sci-fi or sci-fi/moon)
      responses:
        "200":
          description: Tagged books
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  sub_tags:
                    type: array
                    items:
                      type: string
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/ShelfDetailBook'
...
  /me/books/{olId}/tags:
    get:
      summary: Get all label assignments for a book
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - name: olId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookTagResponse'
  /me/books/{olId}/tags/{keyId}:
    put:
      summary: Set a label on a book
      description: |
        For select_one keys, replaces any existing value.
        For select_multiple keys, adds the value (idempotent).
        Use value_id for existing values or value_name to find-or-create.
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - name: olId
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value_id:
                  type: string
                  format: uuid
                value_name:
                  type: string
                  description: "Find-or-create a value by name"
      responses:
        "200":
          description: Tag set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
    delete:
      summary: Remove all label assignments for a key from a book
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - name: olId
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tag unset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...